Support getting UVM metrics by querying the sandbox ID
This is effectively a temporary solution until we put a better long-term
fix in place for querying UVM metrics. This change extends the CRI
stats support to also query a pod sandbox ID. In this case, we look for
the specific UVM metrics exposed by hcsshim, and return those as if they
were the container stats. For the list container stats call, the sandbox
can be specified as either the container ID, pod ID (same as the
container ID), or using the labels for selection.

As part of this, the non-list version of the container stats call has
been changed to just call the list variant, as this was the easiest way
to allow it to check for both a container or a sandbox matching the
given ID.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>